# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the BakingUp Backend API.",
        "title": "BakingUp Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/noti/createNotification": {
            "post": {
                "description": "Create a new notification by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Create a new notification",
                "parameters": [
                    {
                        "description": "Notification Item",
                        "name": "notification_item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateNotificationItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully add a new notification.",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    },
                    "400": {
                        "description": "Cannot add a new notification.",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/noti/deleteNotification": {
            "delete": {
                "description": "Delete a notification by notification id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Delete a notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Noti ID",
                        "name": "noti_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully delete a notification.",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    },
                    "400": {
                        "description": "Cannot delete a notification.",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/noti/getAllNotifications": {
            "get": {
                "description": "Get all notifications by user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Get all notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/domain.NotificationList"
                        }
                    },
                    "400": {
                        "description": "Cannot get all notifications of the user.",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/noti/readAllNotifications": {
            "put": {
                "description": "Read all notification messages by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Read all notification messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully update the read status of all the notifications.",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    },
                    "400": {
                        "description": "Cannot update the read status of all the notifications.",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/noti/readNotification": {
            "put": {
                "description": "Read a notification message by notification id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Read a notification message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Noti ID",
                        "name": "noti_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully update the read status of the notification.",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    },
                    "400": {
                        "description": "Cannot update the read status of the notification.",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/recipe/deleteRecipe": {
            "delete": {
                "description": "Delete a recipe by using recipe id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Delete a recipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "recipe_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    },
                    "400": {
                        "description": "Cannot delete a recipe",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/recipe/getAllRecipes": {
            "get": {
                "description": "Get all recipes by user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Get all recipes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/domain.RecipeList"
                        }
                    },
                    "400": {
                        "description": "Cannot get all recipes",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/recipe/getRecipeDetail": {
            "get": {
                "description": "Get recipe details by recipe ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Get recipe details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "recipe_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/domain.RecipeDetail"
                        }
                    },
                    "400": {
                        "description": "Cannot get recipe detail",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateNotificationItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "eng_message": {
                    "type": "string"
                },
                "eng_title": {
                    "type": "string"
                },
                "is_read": {
                    "type": "boolean"
                },
                "item_id": {
                    "type": "string"
                },
                "item_name": {
                    "type": "string"
                },
                "noti_item_type": {
                    "type": "string"
                },
                "noti_type": {
                    "type": "string"
                },
                "thai_message": {
                    "type": "string"
                },
                "thai_title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.NotificationItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "is_read": {
                    "type": "boolean"
                },
                "item_id": {
                    "type": "string"
                },
                "item_name": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "noti_id": {
                    "type": "string"
                },
                "noti_item_type": {
                    "type": "string"
                },
                "noti_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "domain.NotificationList": {
            "type": "object",
            "properties": {
                "notis": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.NotificationItem"
                    }
                }
            }
        },
...
        "domain.Recipe": {
            "type": "object",
            "properties": {
                "num_of_order": {
                    "type": "integer"
                },
                "recipe_id": {
                    "type": "string"
                },
                "recipe_img": {
                    "type": "string"
                },
                "recipe_name": {
                    "type": "string"
                },
                "servings": {
                    "type": "integer"
                },
                "stars": {
                    "type": "integer"
                },
                "total_time": {
                    "type": "string"
                }
            }
        },
        "domain.RecipeDetail": {
            "type": "object",
            "properties": {
                "instruction_steps": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "instruction_url": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "num_of_order": {
                    "type": "integer"
                },
                "recipe_ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.RecipeIngredient"
                    }
                },
                "recipe_name": {
                    "type": "string"
                },
                "recipe_url": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "servings": {
                    "type": "integer"
                },
                "stars": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "total_time": {
                    "type": "string"
                }
            }
        },
        "domain.RecipeIngredient": {
            "type": "object",
            "properties": {
                "ingredient_name": {
                    "type": "string"
                },
                "ingredient_price": {
                    "type": "number"
                },
                "ingredient_quantity": {
                    "type": "string"
                },
                "ingredient_url": {
                    "type": "string"
                }
            }
        },
        "domain.RecipeList": {
            "type": "object",
            "properties": {
                "recipes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Recipe"
                    }
                }
            }
        },
...
        "http.response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        }
    }
}